requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
huggingface-hub==1.2.3
pillow==12.0.0
pydantic==2.12.5
python-multipart==0.0.21
python-dotenv==1.0.1
